+++ src/withPersistent.ts
@@ -1,7 +1,61 @@
-import { StoreWritable } from 'effector';
+import { Store, StoreWritable } from 'effector';
 import { StoreDriverSingle } from '@cubux/storage-driver';
+import { OptionsWithoutWakeUp, OptionsWithWakeUp } from './internal';
 import { initialize } from './lib/initialize';
 import { WithPersistentOptions } from './types';
+
+interface WithPersistentFn {
+  /**
+   * Register persistent data handling for the given store with the given driver.
+   * @param store Store to work with
+   * @param driver Persistent storage driver
+   * @param key Key for item in driver storage
+   * @param options Other options
+   * @return Input store
+   */
+  <Key, Value, Serialized = Value>(
+    store: StoreWritable<Value>,
+    driver:
+      | StoreDriverSingle<Key, Serialized>
+      | Promise<StoreDriverSingle<Key, Serialized>>,
+    key: Key,
+    options: OptionsWithWakeUp<Value, Value, Serialized>
+  ): typeof store;
+
+  /**
+   * Register persistent data handling for the given store with the given driver.
+   * @param store Store to work with
+   * @param driver Persistent storage driver
+   * @param key Key for item in driver storage
+   * @param options Other options
+   * @return Input store
+   */
+  <Key, Value, Serialized = Value>(
+    store: Store<Value>,
+    driver:
+      | StoreDriverSingle<Key, Serialized>
+      | Promise<StoreDriverSingle<Key, Serialized>>,
+    key: Key,
+    options: OptionsWithWakeUp<Value, Value, Serialized>
+  ): typeof store;
+
+  /**
+   * Register persistent data handling for the given store with the given driver.
+   * @param store Store to work with
+   * @param driver Persistent storage driver
+   * @param key Key for item in driver storage
+   * @param options Other options
+   * @return Input store
+   */
+  <Key, Value, Serialized = Value>(
+    store: StoreWritable<Value>,
+    driver:
+      | StoreDriverSingle<Key, Serialized>
+      | Promise<StoreDriverSingle<Key, Serialized>>,
+    key: Key,
+    options?: OptionsWithoutWakeUp<Value, Value, Serialized>
+  ): typeof store;
+}
 
 /**
  * Register persistent data handling for the given store with the given driver.
@@ -11,7 +65,7 @@ import { WithPersistentOptions } from './types';
  * @param options Other options
  * @return Input store
  */
-export function withPersistent<Key, Value, Serialized = Value>(
+function withPersistentFn<Key, Value, Serialized = Value>(
   store: StoreWritable<Value>,
   driver:
     | StoreDriverSingle<Key, Serialized>
@@ -33,3 +87,13 @@ export function withPersistent<Key, Value, Serialized = Value>(
 
   return store;
 }
+
+/**
+ * Register persistent data handling for the given store with the given driver.
+ * @param store Store to work with
+ * @param driver Persistent storage driver
+ * @param key Key for item in driver storage
+ * @param options Other options
+ * @return Input store
+ */
+export const withPersistent = withPersistentFn as WithPersistentFn;